Update setup.py
I improved the strategy used for prompting and web scraping, and I added some basic error handling. The only remaining error I have noticed, which occurs very infrequently, is that the driver is sometimes not able to click a seemingly valid link. I will decide if it is worth fixing this in the future.